Initial commit of The Funhouse
This is the initial commit of a game I created quite a while ago called
`The Funhouse`.

It was originally written in Visual Basic 4, and this version works with
Visual Basic 5.

I think I tried to import this version in to newer releases of Visual
Basic at some point, and some settings were lost, mostly proper sizing
for the form objects so they fit the room images properly.

Otherwise it seems to work as intended.